# src/main/java/com/misrobot/workflow/controller/ModelController.java
package com.misrobot.workflow.controller;

import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.io.UnsupportedEncodingException;
import java.util.List;

import javax.validation.Valid;
import javax.xml.stream.XMLInputFactory;
import javax.xml.stream.XMLStreamException;
import javax.xml.stream.XMLStreamReader;

import org.activiti.bpmn.converter.BpmnXMLConverter;
import org.activiti.bpmn.model.BpmnModel;
import org.activiti.editor.constants.ModelDataJsonConstants;
import org.activiti.editor.language.json.converter.BpmnJsonConverter;
import org.activiti.engine.ProcessEngine;
import org.activiti.engine.RepositoryService;
import org.activiti.engine.repository.Deployment;
import org.activiti.engine.repository.DeploymentBuilder;
import org.activiti.engine.repository.Model;
import org.activiti.engine.repository.ProcessDefinition;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.util.StringUtils;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.node.ObjectNode;
import com.google.common.collect.Lists;
import com.misrobot.workflow.common.BeanUtils;
import com.misrobot.workflow.controller.request.DeleteModelRequest;
import com.misrobot.workflow.controller.request.DeployModelRequest;
import com.misrobot.workflow.controller.request.NewModelRequest;
import com.misrobot.workflow.controller.response.ModelResponse;
import com.misrobot.workflow.exception.ErrorCode;
import com.misrobot.workflow.exception.WorkflowException;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;

/**
 * 流程模型Model操作相关
 * Created by chenhai on 2017/5/23.
 */
@Api(description = "流程模型Model操作相关", tags = {"model"})
@RestController
@RequestMapping("model")
public class ModelController {

    @Autowired
    private ProcessEngine processEngine;
    @Autowired
    private ObjectMapper objectMapper;
    @Autowired
    private RepositoryService repositoryService;

    /**
     * 新建一个空模型
     *
     * @return
     * @throws UnsupportedEncodingException
     */
    @ApiOperation(value = "新建一个模型")
    @PostMapping
    public ModelResponse newModel(NewModelRequest request) throws UnsupportedEncodingException {
        RepositoryService repositoryService = processEngine.getRepositoryService();
        //初始化一个空模型
        Model model = repositoryService.newModel();

        String name = "new-process";
        if (StringUtils.hasText(request.getName())) {
			name = request.getName();
		}
        String description = "";
        int revision = 1;
        String key = "process";

        ObjectNode modelNode = objectMapper.createObjectNode();
        modelNode.put(ModelDataJsonConstants.MODEL_NAME, name);
        modelNode.put(ModelDataJsonConstants.MODEL_DESCRIPTION, description);
        modelNode.put(ModelDataJsonConstants.MODEL_REVISION, revision);

        model.setName(name);
        model.setKey(key);
        model.setMetaInfo(modelNode.toString());
        model.setTenantId(request.getSystemCode());

        repositoryService.saveModel(model);
        String id = model.getId();

        //完善ModelEditorSource
        ObjectNode editorNode = objectMapper.createObjectNode();
        editorNode.put("id", "canvas");
        editorNode.put("resourceId", "canvas");
        ObjectNode stencilSetNode = objectMapper.createObjectNode();
        stencilSetNode.put("namespace",
                "http://b3mn.org/stencilset/bpmn2.0#");
        editorNode.set("stencilset", stencilSetNode);
        repositoryService.addModelEditorSource(id, editorNode.toString().getBytes("utf-8"));
//        return new ModelAndView("redirect:/modeler.html?modelId=" + id);
        
        ModelResponse modelResponse = new ModelResponse();
        BeanUtils.copyProperties(model, modelResponse);
        return modelResponse;
    }

    /**
     * 获取所有模型
     *
     * @return
     */
    @ApiOperation(value = "获取所有模型")
    @GetMapping
    public List<ModelResponse> modelList() {
        RepositoryService repositoryService = processEngine.getRepositoryService();
        List<ModelResponse> modelResponses = Lists.newArrayList();

        repositoryService.createModelQuery().orderByCreateTime().desc().list().forEach(model -> {
        	ModelResponse modelResponse = new ModelResponse();
        	BeanUtils.copyProperties(model, modelResponse);
        	modelResponses.add(modelResponse);
        });
        
        return modelResponses;
    }

    /**
     * 删除模型
     *
     * @param modelId
     * @return
     * @throws WorkflowException 
     */
    @ApiOperation(value = "删除模型")
    @DeleteMapping("{id}")
    public void deleteModel(@Valid DeleteModelRequest request) throws WorkflowException {
    	Model model = repositoryService.createModelQuery().modelId(request.getModelId())
    			.modelTenantId(request.getSystemCode()).singleResult();
    	if (model == null) {
			throw new WorkflowException(ErrorCode.CANNOT_FIND_MODEL_RECORD, request.getModelId(), request.getSystemCode());
		}
        repositoryService.deleteModel(request.getModelId());
    }

    /**
     * 发布模型为流程定义
     *
     * @param id
     * @return
     * @throws IOException 
     * @throws WorkflowException 
     * @throws Exception
     */
    @ApiOperation(value = "发布模型为流程定义")
    @PostMapping("deployment")
    public void deploy(@Valid DeployModelRequest request) throws WorkflowException {
    	
        //获取模型
        RepositoryService repositoryService = processEngine.getRepositoryService();
        Model modelData = repositoryService.getModel(request.getModelId());
        byte[] bytes = repositoryService.getModelEditorSource(modelData.getId());

        if (bytes == null) {
            throw new WorkflowException(ErrorCode.MODEL_DATA_NULL);
        }

        JsonNode modelNode = null;
		try {
			modelNode = new ObjectMapper().readTree(bytes);
		} catch (IOException e) {
			 throw new WorkflowException(ErrorCode.MODEL_DATA_NULL);
		}

        BpmnModel model = new BpmnJsonConverter().convertToBpmnModel(modelNode);
        if (model.getProcesses().size() == 0) {
        	throw new WorkflowException(ErrorCode.DATA_MODEL_ERROR);
        }
        byte[] bpmnBytes = new BpmnXMLConverter().convertToXML(model);

        //发布流程
        String processName = modelData.getName();
		if (!processName.endsWith(".bpmn20.xml")) {
			processName = processName + ".bpmn20.xml";
		}
        
        DeploymentBuilder deploymentBuilder;
		try {
			deploymentBuilder = repositoryService.createDeployment()
			        .name(modelData.getName())
			        .tenantId(request.getSystemCode())
			        .addString(processName, new String(bpmnBytes, "UTF-8"));
			Deployment deployment = deploymentBuilder.deploy();
			modelData.setDeploymentId(deployment.getId());
			repositoryService.saveModel(modelData);
		} catch (UnsupportedEncodingException e) {
			// ignore
		}
    }

//    @ApiOperation(value = "上传一个已有模型")
//    @RequestMapping(value = "/uploadFile", method = RequestMethod.POST)
//    public void deployUploadedFile(
//            @RequestParam("uploadfile") MultipartFile uploadfile) {
//        InputStreamReader in = null;
//        try {
//            try {
//                String fileName = uploadfile.getOriginalFilename();
//                if (fileName.endsWith(".bpmn20.xml") || fileName.endsWith(".bpmn")) {
//                    XMLInputFactory xif = XMLInputFactory.newInstance();
//                    in = new InputStreamReader(new ByteArrayInputStream(uploadfile.getBytes()), "UTF-8");
//                    XMLStreamReader xtr = xif.createXMLStreamReader(in);
//                    BpmnModel bpmnModel = new BpmnXMLConverter().convertToBpmnModel(xtr);
//
//                    if (bpmnModel.getMainProcess() == null || bpmnModel.getMainProcess().getId() == null) {
////                        notificationManager.showErrorNotification(Messages.MODEL_IMPORT_FAILED,
////                                i18nManager.getMessage(Messages.MODEL_IMPORT_INVALID_BPMN_EXPLANATION));
//                        System.out.println("err1");
//                    } else {
//
//                        if (bpmnModel.getLocationMap().isEmpty()) {
////                            notificationManager.showErrorNotification(Messages.MODEL_IMPORT_INVALID_BPMNDI,
////                                    i18nManager.getMessage(Messages.MODEL_IMPORT_INVALID_BPMNDI_EXPLANATION));
//                            System.out.println("err2");
//                        } else {
//
//                            String processName = null;
//                            if (StringUtils.isNotEmpty(bpmnModel.getMainProcess().getName())) {
//                                processName = bpmnModel.getMainProcess().getName();
//                            } else {
//                                processName = bpmnModel.getMainProcess().getId();
//                            }
//                            Model modelData;
//                            modelData = repositoryService.newModel();
//                            ObjectNode modelObjectNode = new ObjectMapper().createObjectNode();
//                            modelObjectNode.put(ModelDataJsonConstants.MODEL_NAME, processName);
//                            modelObjectNode.put(ModelDataJsonConstants.MODEL_REVISION, 1);
//                            modelData.setMetaInfo(modelObjectNode.toString());
//                            modelData.setName(processName);
//
//                            repositoryService.saveModel(modelData);
//
//                            BpmnJsonConverter jsonConverter = new BpmnJsonConverter();
//                            ObjectNode editorNode = jsonConverter.convertToJson(bpmnModel);
//
//                            repositoryService.addModelEditorSource(modelData.getId(), editorNode.toString().getBytes("utf-8"));
//                        }
//                    }
//                } else {
////                    notificationManager.showErrorNotification(Messages.MODEL_IMPORT_INVALID_FILE,
////                            i18nManager.getMessage(Messages.MODEL_IMPORT_INVALID_FILE_EXPLANATION));
//                    System.out.println("err3");
//                }
//            } catch (Exception e) {
//                String errorMsg = e.getMessage().replace(System.getProperty("line.separator"), "<br/>");
////                notificationManager.showErrorNotification(Messages.MODEL_IMPORT_FAILED, errorMsg);
//                System.out.println("err4");
//            }
//        } finally {
//            if (in != null) {
//                try {
//                    in.close();
//                } catch (IOException e) {
////                    notificationManager.showErrorNotification("Server-side error", e.getMessage());
//                    System.out.println("err5");
//                }
//            }
//        }
//    }

    @PostMapping(value = "convertToModel/{processDefinitionId}")
    public ModelResponse convertToModel(@PathVariable("processDefinitionId") String processDefinitionId)
            throws UnsupportedEncodingException, XMLStreamException {
        ProcessDefinition processDefinition = repositoryService.createProcessDefinitionQuery()
                .processDefinitionId(processDefinitionId).singleResult();
        InputStream bpmnStream = repositoryService.getResourceAsStream(processDefinition.getDeploymentId(),
                processDefinition.getResourceName());
        XMLInputFactory xif = XMLInputFactory.newInstance();
        InputStreamReader in = new InputStreamReader(bpmnStream, "UTF-8");
        XMLStreamReader xtr = xif.createXMLStreamReader(in);
        BpmnModel bpmnModel = new BpmnXMLConverter().convertToBpmnModel(xtr);

        BpmnJsonConverter converter = new BpmnJsonConverter();
        ObjectNode modelNode = converter.convertToJson(bpmnModel);
        Model modelData = repositoryService.newModel();
        modelData.setKey(processDefinition.getKey());
        modelData.setName(processDefinition.getResourceName());
        modelData.setCategory(processDefinition.getDeploymentId());

        ObjectNode modelObjectNode = new ObjectMapper().createObjectNode();
        modelObjectNode.put(ModelDataJsonConstants.MODEL_NAME, processDefinition.getName());
        modelObjectNode.put(ModelDataJsonConstants.MODEL_REVISION, 1);
        modelObjectNode.put(ModelDataJsonConstants.MODEL_DESCRIPTION, processDefinition.getDescription());
        modelData.setMetaInfo(modelObjectNode.toString());

        repositoryService.saveModel(modelData);

        repositoryService.addModelEditorSource(modelData.getId(), modelNode.toString().getBytes("utf-8"));
        
        ModelResponse modelResponse = new ModelResponse();
        BeanUtils.copyProperties(modelData, modelResponse);
        
        return modelResponse;
    }
    
}