# src/main/resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "工作流相关API文档"
    },
    "host": "localhost:8081",
    "basePath": "/workflow",
...
    "paths": {
        "/model": {
            "get": {
                "tags": [
                    "model"
                ],
                "summary": "获取所有模型",
                "operationId": "modelListUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "model"
                ],
                "summary": "新建一个模型",
                "operationId": "newModelUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/model/convertToModel/{processDefinitionId}": {
            "post": {
                "tags": [
                    "model"
                ],
                "summary": "convertToModel",
                "operationId": "convertToModelUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "processDefinitionId",
                        "in": "path",
                        "description": "processDefinitionId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/model/deployment": {
            "post": {
                "tags": [
                    "model"
                ],
                "summary": "发布模型为流程定义",
                "operationId": "deployUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "modelId",
                        "in": "query",
                        "description": "模型id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "systemCode",
                        "in": "query",
                        "description": "系统编码",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/model/{id}": {
            "delete": {
                "tags": [
                    "model"
                ],
                "summary": "删除模型",
                "operationId": "deleteModelUsingDELETE",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "modelId",
                        "in": "query",
                        "description": "模型id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "systemCode",
                        "in": "query",
                        "description": "系统编码",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Model": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string",
                    "format": "date-time"
                },
                "deploymentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "lastUpdateTime": {
                    "type": "string",
                    "format": "date-time"
                },
                "metaInfo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                },
                "version": {
                    "type": "integer",
                    "format": "int32"
                }
            },
            "title": "Model"
        },
...
    }
}